requirements.txt
bcrypt==4.1.3
blinker==1.8.2
bson==0.5.10
click==8.1.7
dnspython==2.6.1
Flask==3.0.3
Flask-JWT-Extended==4.6.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
PyJWT==2.8.0
pymongo==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
Werkzeug==3.0.3
